pyproject.toml
[project]
name = "posprinter-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "escpos>=2.0.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pyserial>=3.5",
    "pywin32>=311",
]
